删除判断+上下架

+++ mingrui-shop-parent/mingrui-shop-service/mingrui-shop-service-xxx/src/main/java/com/baidu/shop/service/impl/BrandServiceImpl.java
@@ -5,8 +5,10 @@ import com.baidu.shop.base.Result;
 import com.baidu.shop.dto.BrandDTO;
 import com.baidu.shop.entity.BrandEntity;
 import com.baidu.shop.entity.CategoryBrandEntity;
+import com.baidu.shop.entity.SpuEntity;
 import com.baidu.shop.mapper.BrandMapper;
 import com.baidu.shop.mapper.CategoryBrandMapper;
+import com.baidu.shop.mapper.SpuMapper;
 import com.baidu.shop.service.BrandService;
 import com.baidu.shop.utils.BaiduBeanUtil;
 import com.baidu.shop.utils.ObjectUtil;
@@ -40,11 +42,14 @@ public class BrandServiceImpl extends BaseApiService implements BrandService {
     @Resource
     private CategoryBrandMapper categoryBrandMapper;
 
+    @Resource
+    private SpuMapper spuMapper;
+
 
     @Override
     public Result<List<BrandEntity>> getBrandByCategory(Integer cid) {
 
-        if(ObjectUtil.isNotNull(cid)){
+        if (ObjectUtil.isNotNull(cid)) {
 
             List<BrandEntity> list = brandMapper.getBrandByCategory(cid);
 
@@ -128,6 +133,15 @@ public class BrandServiceImpl extends BaseApiService implements BrandService {
 
     @Override
     public Result<JsonObject> deleteBrand(Integer id) {
+
+        Example example = new Example(SpuEntity.class);
+        example.createCriteria().andEqualTo("brandId", id);
+
+        List<SpuEntity> spuEntities = spuMapper.selectByExample(example);
+        if (spuEntities.size() >= 1) {
+            return this.setResultError("该品牌绑定有商品,不能删除");
+        }
+
 
         //删除品牌
         brandMapper.deleteByPrimaryKey(id);
